replaces logger config into separate file
